feat: add support for selecting a signing algo for RSA CA keys
The newer defaults for OpenSSH seem to prevent usage of RSA certs with
the default signing algorithm. This patch allows user to request a specific
signing algo.

The server will ask the ssh-agent to sign with a specific algo. This implies
that the server's ssh-agent impl must support the extension flag described
in the draft-miller-ssh-agent-04 Section 4.5.1.