refactor: migrate to yaml.v3
ghodss/yaml is used only in config loading, and the yaml.v3 unmarshaling
appears to be a drop-in replacement; the ghodss/yaml feature of reusing
`json` struct tags for YAML is not used there.

For other uses in the tree, yaml.v3 looks like a drop in replacement for
yaml.v2, too.